backend/requirements.txt
Django==4.2.13
mssql-django==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pyodbc==5.0.1
tzdata==2024.1

